Show task information
I created a python script along with a couple of html templates
in order to allow the web interface to dynamically show task
information through the use of pyrebase, a firebase wrapper library
for python, and flask. I installed pyrebase locally, so if there's
any issues with using this package when attemping to run the script,
that might be why. I haven't tested for this yet, since I wanted to
get it up and running in my development environment first.

The ultimate goal is to move the script and other project web files
into wherever we need to put them to make the whole GUI live through
the firebase hosting system, but for now, running the script allows
you to view the pages via localhost:5000 since I don't know how to
connect a python script to the firebase yet. If anyone can teach me
or knows a resource I can use to teach myself how to do it, I'm all
ears.

The code isn't broken, but there is a known bug here where the
information given to the template by the calling page is incorrect.
For the task name and description, "none" is passed into the page.
For the task owners and commits, nothing shows, but the list
framework appears in the source page on the browser, which means
that flask sees that something exists in the variable, but it comes
up empty when running through the loops. The specific task that's
being called through "hello.html" is one where all four sections
have data stored in them, so we should be seeing that data instead.

I will continue to hack on this bug in an effort to show the
correct information, but I wanted to provide an initial template
example for Travis and Isaiah to work with, who are the other
developers working with me on the GUI.

Below is my idea for how we tell our taskmanager to create a new
task. It would probably be at the very end of the commit whenever
any taskmanager business showed up. However, this is only an idea,
not set in stone, so feel free to design for a different format as
you see fit, cloud functions team.

TASKMANAGER: NEW
	--name: show task info
	--description: Information shown is incorrect (empty). Needs to show information from database related to a given task.
	--owners: Susan