Show task info (cont'd)
TASK: Show task info
ACTION: Finished writing tasks page. Now listens continuously for
changes to a given task in the firebase while on the page. Also
made to be able to show only information that exists in the
firebase document for that task. Is able to show up to all of the
following: task name; due date; the date of last update; task
description; a list of all owners of the task; and commits, which
are ordered by commit key -- the date of the commit -- from the
earliest to the most recent. Sadly, firebase doesn't support
descending ordering. I looked.

I'm also working on highlighting the due date based on the date of
the last action, but for whatever reason, the DOM call to change
the background color doesn't run, even though my loop and function
run as expected. It's not a necessary thing for this page, but
the functionality might be useful for both the task page and the
kanban page as a sort of glitter somewhere on down the line. If
anyone has the time, please give it a glance and see if maybe I
made a syntax or capitalization error because the DOM call should
run and it doesn't.