Add and remove users
Wrote javascript functions to add and remove users to both a
specific task and a specific board, using the information stored
in sessionStorage. See "public/task.html" for a basic functionality
demo. All four of these functions should in practice be called
through a popup or other appearing form, so we're not allowing
end-users to add and remove users from the boards and tasks like
they're god or something. Haha.

Additionally, I moved the giant wall of javascript from the head of
"public/task.html" into "scripts/functions.js" since it made more
sense to organize it that way, overwriting what was there in favor
of code that's been proven to work. Plus the new functions described
above.