Adding database functionality to kanban
Made every session determined on the user logged in so that they see
their boards and task for each respective board. Made some changes to
css.

Known bugs - clicking buttons for adding and removing users or owners
are not fully implemented therefore may cause errors if clicked on.
Clicking on a task link does not lead to the task page at the moment, but is being
worked on.