Implement HTML drag and drop
Created a javascript file with everything you need to drag tasks
from one column to another. Also implemented database interaction
to reflect task movement. This code should be bug-free with regards
to dropping onto an element that has multiple draggable things in
it and moving a task from one column to another in the database.

Things to note:

	1 - If the task doesn't exist at the location specified via
the drag and drop function calls, nothing happens but the console
does print out an error message telling you so.

	2 - If you need to see how to implement these functions,
ask me and I can send you the html file I used for local testing.
I'm not putting it in the repository because it's not going to be
used as part of the final solution.

	3 - These functions are designed specifically for use in
"kanban.html". There's no need to link this script file into any
other pages on the site.

If you have any questions about this commit, let me know and I'll
respond as soon as I can.