Added more database functionality to kanban page and cleaned up CSS
- changed as much as possible of CSS to be based on percentage values
rather than hard coded pixels so that no matter what the size of the
screen is it the page format should always look the same
- added a drop down list of users as opposed to using a textbox to type in a user to
add or remov users from a board. Made the add list and drop list dynamic
to each page value and based the list of users off of who is already a
user and who isn't. Also fixed an error that was causing the list of
usernames to duplicate
- connected Susan's add/remove user to the board functions to working
buttons
- set permissions so that only owners could see the "add" and "remove"
owner buttons
- fixed an error that was displaying every child of a column rather than
only displaying the 'tasks' of a column

Bugs currently being worked on:
- drop down list of users below the "add owners" button is not fully
implemnted correctly. The list goes upward rather than downward, the
list may contain duplicate names, the list is not dynamically based off
of who is a owner and who isnt an owner as of yet
- the buttons to add or remove an owner are not fully implemented yet
as right now they do nothing