Add error checking & more db interaction to add/rem user functions
Added some error checking to the add/remove user functions for
tasks and boards. This limits adding and removing users only to
those that exists in the "Users" document of the database. That
being said, adding and removing a user to/from a board now is
reflected in the "Users" document as a true/false boolean (exactly
like how we do it with task owners and board users), where the
board key is the key and the value is either true or false.