Made the users of a board displayed at the top of the page rather than within each individual task to reduce clutter.
- Added commas between user names to clearly seperate each name from one
another
- Added commas between owner names to clearly seperate each name from
one another
- Fixed the drop down list under the "add owners" button to be dynamic
by not showing users who are already an owner and to show a list of
users who are not an owner of said task. Also fixed duplication of
usernames in list (was an error).

NOTE - In a previous push I noted that the drop down list going upward
was a bug, but this is in fact simply how drop down lists work. They go
up or down based on where the center of the page is in comparison to
where the button is on the page.

Current Bugs that I am aware of and working on:

- add/remove buttons do not seem to do anything (not even display an
error in the console) on the firefox browser... Something in our code is
not supported by firefox maybe? Microsoft edge and google chrome seem to
run with no issues

- add/remove users from a board buttons do not work unless you first
click on a board (left side of the page)

- If a user signs out and a new user signs in the same computer and
browser (without closing the browser tabs), the default kanban view is
that of the previous logged on user

- add/remove owner buttons not working at the moment

- permissions is not 100% accurate as sometimes owner buttons are
displayed to the currently logged in user when in fact they are not the
owner of that task