Fix drag&drop, task.html; create new task
Pretty self-explanatory.

To create a new task, have a button link to "newTask.html" and
this will bring up the form to add a task. It doesn't necessarily
look pretty but it does what it needs to. Pretty can come later.

Updated "task.html" and "functions.js" to write the date for each
commit, as well as display an action history.

I don't remember what I did for drag and drop at this point, but
it was probably some minor bug I found and fixed.

Home stretch, fam. Let's go!