Repushing code that was removed in someone else's push.
That code consists of displaying user names at the top of the page along
with some other minor changes.

In addition to repushing that code, I fixed a few issues:

- Fixed the adding/removing users to a board buttons and how they were not working
unless you had clicked on a board first. (Boards on the left side of the
page)

- I also fixed an issue where the adding/remove users to/from a board
where based on the name of the board. Now they are appropriately based
on the ID of a board.