Minor edits to drag&drop and newTask stuff
For drag&drop and newTask.js, added a little string in front of the
timestamp to help keep things in order in the database.

For newTask.html, made the due date required, changed the button to
type submit, and removed the onclick function call.

By requiring all four of these things on the newTask form, we should
be able to eliminate the need to update a task to add a due date...
At least until later if/when we have time to keep the code up to
date.