Fix drag and drop into original column bug
Isaiah and I found a bug where dragging a task from a column and
dropping it back into its original column would remove and replace
the task, effectively deleting everything except the new changes to
the ActionHistory and LatestAction documents in the database. This
was solved by checking for a column change before calling the
function to move the task.