Last major changes being pushed
- Changed display of owner buttons to be within Task page
- Implemented Drag and Drop within kanban page
- added slidebars for columns if their height or width extends the main
page
- implemented new column button and functionality
- implemented new tasks button and functionality
- did an overhaul of removing code that was not needed in files and made
more use of dynamic coding rather than hardcoded values
- changes CSS a little for a better viewing kanban page

Current bugs aware of:

- adding board button not currently working (will be fixed by tomorrow)

- when in the task page, clicking on add or remove button works, however
the list of users will not be updated from the database unless the page
refresh button on the browser is pressed (Will be fixed by tomorrow)

- when dragging and dropping unwanted space is created after header and
columns. Everything still is functional and performs as needed, this is
more of a visual annoyance than anything. To remove unwanted space,
simply click on the board your currently viewing (list of boards on the
left side of the page) (will not be fixed)